Restructure test directory
This structure more closely aligns with the python best practices
